+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/application.js
@@ -15,7 +15,7 @@ OpsChooserApp = Backbone.Marionette.Application.extend({
 
     set_datasource: function(datasource) {
         $("#datasource .btn[data-value='" + datasource + "']").button('toggle');
-        this.queryBuilderView.cql_field_chooser_setup();
+        this.queryBuilderView.setup_cql_field_chooser();
         this.queryBuilderView.setup_common_form();
         this.queryBuilderView.setup_comfort_form();
     },